@@ -46,7 +46,7 @@ if(COMMAND cmake_policy)
   cmake_policy(SET CMP0023 NEW)
 endif()
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
@@ -151,7 +151,6 @@ if(${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
 endif()
 
 ## Add lots of dependencies to compiler switches
-include(SuperTux/ProvideBoost)
 include(SuperTux/ProvideGlm)
 include(SuperTux/ProvideFmt)
 include(SuperTux/ProvideSDL2)
@@ -264,7 +263,6 @@ target_link_libraries(supertux2_lib PUBLIC LibSavePNG)
 target_link_libraries(supertux2_lib PUBLIC LibPartioZip)
 target_link_libraries(supertux2_lib PUBLIC LibOpenAL)
 target_link_libraries(supertux2_lib PUBLIC LibOggVorbis)
-target_link_libraries(supertux2_lib PUBLIC LibBoost)
 target_link_libraries(supertux2_lib PUBLIC LibGlm)
 target_link_libraries(supertux2_lib PUBLIC LibFmt)
 target_link_libraries(supertux2_lib PUBLIC LibPhysfs)
