@@ -17,7 +17,7 @@
 
 #include "addon/addon.hpp"
 
-#include <boost/optional.hpp>
+#include <optional>
 #include <fmt/format.h>
 #include <sstream>
 
@@ -139,7 +139,7 @@ Addon::parse(const ReaderMapping& mapping)
     mapping.get("url", addon->m_url);
     mapping.get("md5", addon->m_md5);
     mapping.get("format", addon->m_format);
-    boost::optional<ReaderCollection> reader;
+    std::optional<ReaderCollection> reader;
     if (mapping.get("screenshots", reader))
     {
       for (auto& obj : reader->get_objects())
