@@ -29,10 +29,8 @@
 #include "supertux/menu/menu_storage.hpp"
 #include "supertux/screen_manager.hpp"
 #include "supertux/game_manager.hpp"
-#include "supertux/textscroller_screen.hpp"
 #include "supertux/world.hpp"
 #include "util/log.hpp"
-#include "util/file_system.hpp"
 #include "video/video_system.hpp"
 #include "video/viewport.hpp"
 
@@ -50,16 +48,13 @@
 
 MainMenu::MainMenu()
 {
-  set_center_pos(static_cast<float>(SCREEN_WIDTH) / 2.0f,
-                 static_cast<float>(SCREEN_HEIGHT) / 2.0f + 35.0f);
-
-  add_entry(MNID_STARTGAME, _("Start Game"));
+  add_submenu(_("Start Game"), MenuStorage::WORLDSET_MENU);
   // TODO: Manage to build OpenSSL for Emscripten so we can build CURL so we can
   //       build the add-ons so we can re-enable them.
   //       Also see src/addon/downloader.*pp
-  add_entry(MNID_ADDONS, _("Add-ons"));
+  add_submenu(_("Add-ons"), MenuStorage::ADDON_MENU);
 #ifdef __EMSCRIPTEN__
-  add_entry(MNID_MANAGEASSETS, _("Manage Assets"));
+  add_entry(_("Manage Assets"), MenuStorage::ASSET_MENU);
 #endif
   add_submenu(_("Options"), MenuStorage::OPTIONS_MENU);
   add_entry(MNID_LEVELEDITOR, _("Level Editor"));
@@ -71,6 +66,8 @@ MainMenu::MainMenu()
 #ifndef REMOVE_QUIT_BUTTON
   add_entry(MNID_QUITMAINMENU, _("Quit"));
 #endif
+
+  on_window_resize();
 }
 
 void
@@ -85,52 +82,40 @@ MainMenu::menu_action(MenuItem& item)
 {
   switch (item.get_id())
   {
-
-    case MNID_STARTGAME:
-      // World selection menu
-      MenuManager::instance().push_menu(MenuStorage::WORLDSET_MENU);
-      break;
-
-    case MNID_ADDONS:
-      // Add-ons Menu
-      MenuManager::instance().push_menu(MenuStorage::ADDON_MENU);
-      break;
-
-    case MNID_MANAGEASSETS:
-      MenuManager::instance().push_menu(MenuStorage::ASSET_MENU);
-      break;
-
-     case MNID_CREDITS:
+    case MNID_CREDITS:
     {
-      // Credits Level
       SoundManager::current()->stop_music(0.2f);
       std::unique_ptr<World> world = World::from_directory("levels/misc");
-      GameManager::current()->start_level(*world, "credits.stl");
+      GameManager::current()->start_level(*world, "credits.stl"); // Credits Level
+      break;
     }
-	  break;
 
     case MNID_LEVELEDITOR:
-      {
-        MenuManager::instance().clear_menu_stack();
-        std::unique_ptr<Screen> screen(new Editor());
-        auto fade = std::make_unique<FadeToBlack>(FadeToBlack::FADEOUT, 0.5f);
-        SoundManager::current()->stop_music(0.5);
-        ScreenManager::current()->push_screen(std::move(screen), std::move(fade));
-        //Editor::current()->setup();
-      }
+    {
+      MenuManager::instance().clear_menu_stack();
+      std::unique_ptr<Screen> screen(new Editor());
+      auto fade = std::make_unique<FadeToBlack>(FadeToBlack::FADEOUT, 0.5f);
+      SoundManager::current()->stop_music(0.5);
+      ScreenManager::current()->push_screen(std::move(screen), std::move(fade));
+      //Editor::current()->setup();
       break;
+    }
 
     case MNID_DONATE:
+    {
       Dialog::show_confirmation(_("This will take you to the SuperTux donation page. Are you sure you want to continue?"), [] {
         FileSystem::open_url("https://www.supertux.org/donate.html");
       });
       break;
+    }
 
     case MNID_QUITMAINMENU:
+    {
       MenuManager::instance().clear_menu_stack();
       ScreenManager::current()->quit(std::unique_ptr<ScreenFade>(new FadeToBlack(FadeToBlack::FADEOUT, 0.25f)));
       SoundManager::current()->stop_music(0.25);
       break;
+    }
   }
 }
 
