@@ -16,8 +16,7 @@
 
 #include "util/file_system.hpp"
 
-#include <boost/filesystem.hpp>
-#include <boost/version.hpp>
+#include <filesystem>
 #include <sstream>
 #include <stdexcept>
 #include <sys/stat.h>
@@ -40,14 +39,14 @@
 #include "util/log.hpp"
 #include "util/string_util.hpp"
 
-namespace fs = boost::filesystem;
+namespace fs = std::filesystem;
 
 namespace FileSystem {
 
 bool exists(const std::string& path)
 {
   fs::path location(path);
-  boost::system::error_code ec;
+  std::error_code ec;
 
   // If we get an error (such as "Permission denied"), then ignore it
   // and pretend that the path doesn't exist.
@@ -71,11 +70,7 @@ void mkdir(const std::string& directory)
 
 void copy(const std::string& source_path, const std::string& target_path)
 {
-#if BOOST_VERSION < 107400
-  fs::copy_file(source_path, target_path, boost::filesystem::copy_option::overwrite_if_exists);
-#else
-  fs::copy_file(source_path, target_path, boost::filesystem::copy_options::overwrite_existing);
-#endif
+  fs::copy_file(source_path, target_path, fs::copy_options::overwrite_existing);
 }
 
 std::string dirname(const std::string& filename)
@@ -102,43 +97,7 @@ std::string basename(const std::string& filename)
 
 std::string relpath(const std::string& filename, const std::string& basedir)
 {
-#if BOOST_VERSION >= 106000
   return fs::relative(filename, basedir).string();
-#else
-  fs::path from = basedir;
-  fs::path to = filename;
-
-  // Taken from https://stackoverflow.com/a/29221546
-
-  // Start at the root path and while they are the same then do nothing then when they first
-  // diverge take the entire from path, swap it with '..' segments, and then append the remainder of the to path.
-  fs::path::const_iterator fromIter = from.begin();
-  fs::path::const_iterator toIter = to.begin();
-
-  // Loop through both while they are the same to find nearest common directory
-  while (fromIter != from.end() && toIter != to.end() && (*toIter) == (*fromIter))
-  {
-    ++toIter;
-    ++fromIter;
-  }
-
-  // Replace from path segments with '..' (from => nearest common directory)
-  fs::path finalPath;
-  while (fromIter != from.end())
-  {
-    finalPath /= "..";
-    ++fromIter;
-  }
-
-  // Append the remainder of the to path (nearest common directory => to)
-  while (toIter != to.end())
-  {
-    finalPath /= *toIter;
-    ++toIter;
-  }
-
-  return finalPath.string();
-#endif
 }
 
 std::string extension(const std::string& filename)
