@@ -16,8 +16,6 @@
 
 #include "util/reader_mapping.hpp"
 
-#include <boost/ref.hpp>
-#include <boost/utility/typed_in_place_factory.hpp>
 #include <sexp/io.hpp>
 #include <sstream>
 #include <stdexcept>
@@ -80,25 +78,25 @@ ReaderMapping::get_item(const char* key) const
   }
 
 bool
-ReaderMapping::get(const char* key, bool& value, const boost::optional<bool>& default_value) const
+ReaderMapping::get(const char* key, bool& value, const std::optional<bool>& default_value) const
 {
   GET_VALUE_MACRO("bool", is_boolean, as_bool)
 }
 
 bool
-ReaderMapping::get(const char* key, int& value, const boost::optional<int>& default_value) const
+ReaderMapping::get(const char* key, int& value, const std::optional<int>& default_value) const
 {
   GET_VALUE_MACRO("int", is_integer, as_int)
 }
 
 bool
-ReaderMapping::get(const char* key, uint32_t& value, const boost::optional<uint32_t>& default_value) const
+ReaderMapping::get(const char* key, uint32_t& value, const std::optional<uint32_t>& default_value) const
 {
   GET_VALUE_MACRO("uint32_t", is_integer, as_int)
 }
 
 bool
-ReaderMapping::get(const char* key, float& value, const boost::optional<float>& default_value) const
+ReaderMapping::get(const char* key, float& value, const std::optional<float>& default_value) const
 {
   GET_VALUE_MACRO("float", is_real, as_float)
 }
@@ -106,7 +104,7 @@ ReaderMapping::get(const char* key, float& value, const boost::optional<float>& 
 #undef GET_VALUE_MACRO
 
 bool
-ReaderMapping::get(const char* key, std::string& value, const boost::optional<const char*>& default_value) const
+ReaderMapping::get(const char* key, std::string& value, const std::optional<const char*>& default_value) const
 {
   auto const sx = get_item(key);
   if (!sx) {
@@ -159,7 +157,7 @@ ReaderMapping::get(const char* key, std::string& value, const boost::optional<co
 
 bool
 ReaderMapping::get(const char* key, std::vector<bool>& value,
-                   const boost::optional<std::vector<bool>>& default_value) const
+                   const std::optional<std::vector<bool>>& default_value) const
 {
   value.clear();
   GET_VALUES_MACRO("bool", is_boolean, as_bool)
@@ -167,7 +165,7 @@ ReaderMapping::get(const char* key, std::vector<bool>& value,
 
 bool
 ReaderMapping::get(const char* key, std::vector<int>& value,
-                   const boost::optional<std::vector<int>>& default_value) const
+                   const std::optional<std::vector<int>>& default_value) const
 {
   value.clear();
   GET_VALUES_MACRO("int", is_integer, as_int)
@@ -176,7 +174,7 @@ ReaderMapping::get(const char* key, std::vector<int>& value,
 
 bool
 ReaderMapping::get(const char* key, std::vector<float>& value,
-                   const boost::optional<std::vector<float>>& default_value) const
+                   const std::optional<std::vector<float>>& default_value) const
 {
   value.clear();
   GET_VALUES_MACRO("float", is_real, as_float)
@@ -184,7 +182,7 @@ ReaderMapping::get(const char* key, std::vector<float>& value,
 
 bool
 ReaderMapping::get(const char* key, std::vector<std::string>& value,
-                   const boost::optional<std::vector<std::string>>& default_value) const
+                   const std::optional<std::vector<std::string>>& default_value) const
 {
   value.clear();
   GET_VALUES_MACRO("string", is_string, as_string)
@@ -192,7 +190,7 @@ ReaderMapping::get(const char* key, std::vector<std::string>& value,
 
 bool
 ReaderMapping::get(const char* key, std::vector<unsigned int>& value,
-                   const boost::optional<std::vector<unsigned int>>& default_value) const
+                   const std::optional<std::vector<unsigned int>>& default_value) const
 {
   value.clear();
   GET_VALUES_MACRO("unsigned int", is_integer, as_int)
@@ -201,11 +199,11 @@ ReaderMapping::get(const char* key, std::vector<unsigned int>& value,
 #undef GET_VALUES_MACRO
 
 bool
-ReaderMapping::get(const char* key, boost::optional<ReaderMapping>& value) const
+ReaderMapping::get(const char* key, std::optional<ReaderMapping>& value) const
 {
   auto const sx = get_item(key);
   if (sx) {
-    value = boost::in_place<ReaderMapping>(boost::ref(m_doc), boost::ref(*sx));
+    value.emplace(m_doc, *sx);
     return true;
   } else {
     return false;
@@ -213,11 +211,11 @@ ReaderMapping::get(const char* key, boost::optional<ReaderMapping>& value) const
 }
 
 bool
-ReaderMapping::get(const char* key, boost::optional<ReaderCollection>& value) const
+ReaderMapping::get(const char* key, std::optional<ReaderCollection>& value) const
 {
   auto const sx = get_item(key);
   if (sx) {
-    value = boost::in_place<ReaderCollection>(boost::ref(m_doc), boost::ref(*sx));
+    value.emplace(m_doc, *sx);
     return true;
   } else {
     return false;
