@@ -17,6 +17,7 @@
 #include "worldmap/worldmap_parser.hpp"
 
 #include <physfs.h>
+#include <sstream>
 
 #include "object/ambient_light.hpp"
 #include "object/background.hpp"
@@ -81,7 +82,7 @@ WorldMapParser::load_worldmap(const std::string& filename)
       m_worldmap.m_tileset = TileManager::current()->get_tileset("images/ice_world.strf");
     }
 
-    boost::optional<ReaderMapping> sector;
+    std::optional<ReaderMapping> sector;
     if (!level_.get("sector", sector)) {
       throw std::runtime_error("No sector specified in worldmap file.");
     } else {
