@@ -1,5 +1,6 @@
 //  SuperTux
 //  Copyright (C) 2009 Ingo Ruhnke <grumbel@gmail.com>
+//                2023 Vankata453
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -21,231 +22,146 @@
 #include "gui/menu.hpp"
 #include "gui/mousecursor.hpp"
 #include "gui/notification.hpp"
-#include "supertux/gameconfig.hpp"
 #include "supertux/globals.hpp"
 #include "supertux/menu/menu_storage.hpp"
 #include "util/log.hpp"
 #include "video/drawing_context.hpp"
 
-MenuManager* MenuManager::s_instance = nullptr;
+const float MenuManager::s_menu_repeat_initial = 0.4f;
+const float MenuManager::s_menu_repeat_rate = 0.1f;
 
 MenuManager&
 MenuManager::instance()
 {
-  assert(s_instance);
-  return *s_instance;
+  return *current();
 }
 
-namespace {
-
-Rectf menu2rect(const Menu& menu)
-{
-  return Rectf(menu.get_center_pos().x - menu.get_width() / 2,
-               menu.get_center_pos().y - menu.get_height() / 2,
-               menu.get_center_pos().x + menu.get_width() / 2,
-               menu.get_center_pos().y + menu.get_height() / 2);
-}
-
-Rectf dialog2rect(const Dialog& dialog)
-{
-  return Rectf(dialog.get_center_pos().x - dialog.get_width() / 2,
-               dialog.get_center_pos().y - dialog.get_height() / 2,
-               dialog.get_center_pos().x + dialog.get_width() / 2,
-               dialog.get_center_pos().y + dialog.get_height() / 2);
-}
-
-} // namespace
-
-class MenuTransition final
-{
-private:
-  Rectf m_from_rect;
-  Rectf m_to_rect;
-
-  float m_effect_progress;
-  float m_effect_start_time;
-  bool m_is_active;
-
-public:
-  MenuTransition() :
-    m_from_rect(),
-    m_to_rect(),
-    m_effect_progress(1.0f),
-    m_effect_start_time(),
-    m_is_active(false)
-  {
-  }
-
-  void start(const Rectf& from_rect,
-             const Rectf& to_rect)
-  {
-    m_from_rect = from_rect;
-    m_to_rect = to_rect;
-
-    m_effect_start_time = g_real_time;
-    m_effect_progress = 0.0f;
-
-    m_is_active = true;
-  }
-
-  void set(const Rectf& rect)
-  {
-    m_to_rect = m_from_rect = rect;
-  }
-
-  void update()
-  {
-    if (!g_config->transitions_enabled && m_is_active)
-    {
-      m_effect_progress = 1.0f;
-      m_is_active = false;
-      return;
-    }
-    if (m_is_active)
-    {
-      m_effect_progress = (g_real_time - m_effect_start_time) * 6.0f;
-
-      if (m_effect_progress > 1.0f)
-      {
-        m_effect_progress = 1.0f;
-        m_is_active = false;
-      }
-    }
-  }
-
-  void draw(DrawingContext& context)
-  {
-    float p = m_effect_progress;
-
-    Rectf rect = m_to_rect;
-    if (m_is_active)
-    {
-      rect = Rectf((m_to_rect.get_left() * p) + (m_from_rect.get_left() * (1.0f - p)),
-                   (m_to_rect.get_top() * p) + (m_from_rect.get_top() * (1.0f - p)),
-                   (m_to_rect.get_right() * p) + (m_from_rect.get_right() * (1.0f - p)),
-                   (m_to_rect.get_bottom() * p) + (m_from_rect.get_bottom() * (1.0f - p)));
-    }
-
-    // draw menu background rectangles
-    context.color().draw_filled_rect(Rectf(rect.get_left() - 4, rect.get_top() - 10-4,
-                                             rect.get_right() + 4, rect.get_bottom() + 10 + 4),
-                                       g_config->menubackcolor,
-                                       g_config->menuroundness + 4.f,
-                                       LAYER_GUI-10);
-
-    context.color().draw_filled_rect(Rectf(rect.get_left(), rect.get_top() - 10,
-                                             rect.get_right(), rect.get_bottom() + 10),
-                                       g_config->menufrontcolor,
-                                       g_config->menuroundness,
-                                       LAYER_GUI-10);
-  }
-
-  bool is_active() const
-  {
-    return m_is_active;
-  }
-};
 
 MenuManager::MenuManager() :
-  m_dialog(),
-  m_has_next_dialog(false),
-  m_next_dialog(),
-  m_notification(),
-  m_has_next_notification(false),
-  m_next_notification(),
   m_menu_stack(),
-  m_transition(new MenuTransition)
+  m_transition(new MenuTransition),
+  m_menu_repeat_time(),
+  m_dialog(),
+  m_notification()
 {
-  s_instance = this;
 }
 
 MenuManager::~MenuManager()
 {
-  s_instance = nullptr;
 }
 
 void
 MenuManager::refresh()
 {
   for (const auto& menu : m_menu_stack)
-  {
     menu->refresh();
+}
+
+void
+MenuManager::check_input_action(Control control, MenuAction action,
+                                const Controller& controller, MenuAction& result)
+{
+  if (controller.pressed(control))
+  {
+    result = action;
+    m_menu_repeat_time = g_real_time + s_menu_repeat_initial;
+  }
+  if (controller.hold(control) &&
+      m_menu_repeat_time != 0 && g_real_time > m_menu_repeat_time)
+  {
+    result = action;
+    m_menu_repeat_time = g_real_time + s_menu_repeat_rate;
   }
 }
 
 void
 MenuManager::process_input(const Controller& controller)
 {
-  if (m_dialog && !m_dialog->is_passive())
+  if (m_dialog.current && !m_dialog.current->is_passive())
   {
-    m_dialog->process_input(controller);
+    m_dialog.current->process_input(controller);
   }
   else if (current_menu())
   {
-    current_menu()->process_input(controller);
+    MenuAction action = MenuAction::NONE;
+
+    check_input_action(Control::UP, MenuAction::UP, controller, action);
+    check_input_action(Control::DOWN, MenuAction::DOWN, controller, action);
+    check_input_action(Control::LEFT, MenuAction::LEFT, controller, action);
+    check_input_action(Control::RIGHT, MenuAction::RIGHT, controller, action);
+
+    if (controller.pressed(Control::ACTION) ||
+        controller.pressed(Control::JUMP) ||
+        controller.pressed(Control::MENU_SELECT) ||
+        (!current_menu()->is_sensitive() && controller.pressed(Control::MENU_SELECT_SPACE)))
+    {
+      action = MenuAction::HIT;
+    }
+
+    if (controller.pressed(Control::ESCAPE) ||
+        controller.pressed(Control::CHEAT_MENU) ||
+        controller.pressed(Control::DEBUG_MENU) ||
+        controller.pressed(Control::MENU_BACK))
+    {
+      action = MenuAction::BACK;
+    }
+
+    check_input_action(Control::REMOVE, MenuAction::REMOVE, controller, action);
+
+    current_menu()->process_action(action);
   }
 }
 
 void
 MenuManager::event(const SDL_Event& ev)
 {
-  if (!m_transition->is_active())
-  {
-    if (m_dialog && !m_dialog->is_passive())
-    {
-      m_dialog->event(ev);
-    }
-    else if (current_menu())
-    {
-      // only pass events when the menu is fully visible and not in a
-      // transition animation
-      current_menu()->event(ev);
-    }
+  if (m_notification.current)
+    m_notification.current->event(ev);
 
-    if (m_notification)
-    {
-      m_notification->event(ev);
-    }
-  }
+  if (m_transition->is_active()) return; // Do not continue, if a transition is active
+
+  if (m_dialog.current && !m_dialog.current->is_passive())
+    m_dialog.current->event(ev);
+  else if (current_menu())
+    current_menu()->event(ev);
 }
 
 void
 MenuManager::draw(DrawingContext& context)
 {
-  if (m_has_next_dialog)
+  if (m_dialog.has_next) // Has next dialog
   {
-    if (m_next_dialog != nullptr) m_next_dialog->update();
-    if (m_dialog != nullptr && m_next_dialog != nullptr)
+    if (m_dialog.next) m_dialog.next->update();
+
+    if (m_dialog.current && m_dialog.next)
     {
-      if (!m_dialog->is_passive() && !m_next_dialog->is_passive())
-      {
-        transition(m_dialog.get(), m_next_dialog.get());
-      }
-      else if (!m_next_dialog->is_passive())
-      {
-        transition(current_menu(), m_next_dialog.get());
-      }
-      else if (!m_dialog->is_passive())
-      {
-        transition(m_dialog.get(), current_menu());
-      }
+      const bool current_passive = m_dialog.current->is_passive();
+      const bool next_passive = m_dialog.next->is_passive();
+
+      if (!current_passive && !next_passive)
+        transition(m_dialog.current.get(), m_dialog.next.get());
+      else if (current_passive)
+        transition(current_menu(), m_dialog.next.get());
+      else if (next_passive)
+        transition(m_dialog.current.get(), current_menu());
     }
-    else if (m_dialog != nullptr)
+    else if (m_dialog.current && !m_dialog.current->is_passive())
     {
-      if (!m_dialog->is_passive()) transition(m_dialog.get(), current_menu());
+      transition(m_dialog.current.get(), current_menu());
     }
-    else if (m_next_dialog != nullptr)
+    else if (m_dialog.next && !m_dialog.next->is_passive())
     {
-      if (!m_next_dialog->is_passive()) transition(current_menu(), m_next_dialog.get());
+      transition(current_menu(), m_dialog.next.get());
     }
-    m_dialog = std::move(m_next_dialog);
-    m_has_next_dialog = false;
+
+    m_dialog.current = std::move(m_dialog.next);
+    m_dialog.has_next = false;
   }
 
-  if (m_has_next_notification)
+  if (m_notification.has_next) // Has next notification
   {
-    m_notification = std::move(m_next_notification);
-    m_has_next_notification = false;
+    m_notification.current = std::move(m_notification.next);
+    m_notification.has_next = false;
   }
 
   if (m_transition->is_active())
@@ -253,51 +169,73 @@ MenuManager::draw(DrawingContext& context)
     m_transition->update();
     m_transition->draw(context);
   }
-  else
+  else // Has no active transition
   {
-    if (m_dialog)
+    if (m_dialog.current)
     {
-      m_dialog->update();
-      m_dialog->draw(context);
+      m_dialog.current->update();
+      m_dialog.current->draw(context);
     }
-    if (current_menu() && (!m_dialog || m_dialog->is_passive()))
+
+    if (current_menu() && (!m_dialog.current || m_dialog.current->is_passive()))
     {
       // brute force the transition into the right shape in case the
       // menu has changed sizes
-      m_transition->set(menu2rect(*current_menu()));
+      m_transition->set(to_rect(*current_menu()));
       m_transition->draw(context);
 
       current_menu()->draw(context);
     }
   }
 
-  if (m_notification)
-  {
-    m_notification->draw(context);
-  }
+  if (m_notification.current) // Has current notification
+    m_notification.current->draw(context);
 
-  if ((m_dialog || current_menu()) && MouseCursor::current())
-  {
+  if ((has_dialog() || is_active()) && MouseCursor::current()) // Cursor should be drawn
     MouseCursor::current()->draw(context);
-  }
 }
+
 
 void
 MenuManager::set_dialog(std::unique_ptr<Dialog> dialog)
 {
-  // delay reseting m_dialog to a later point, as otherwise the Dialog
-  // can't unset itself without ending up with "delete this" problems
-  m_next_dialog = std::move(dialog);
-  m_has_next_dialog = true;
+  m_dialog.next = std::move(dialog);
+  m_dialog.has_next = true;
 }
 
 void
 MenuManager::set_notification(std::unique_ptr<Notification> notification)
 {
-  // delay reseting m_notification to a later point, as otherwise the Notification
-  // can't unset itself without ending up with "delete this" problems
-  m_next_notification = std::move(notification);
-  m_has_next_notification = true;
+  m_notification.next = std::move(notification);
+  m_notification.has_next = true;
+}
+
+
+void
+MenuManager::set_menu(int id, bool skip_transition)
+{
+  set_menu(MenuStorage::instance().create(static_cast<MenuStorage::MenuId>(id)), skip_transition);
+}
+
+void
+MenuManager::set_menu(std::unique_ptr<Menu> menu, bool skip_transition)
+{
+  if (menu)
+  {
+    if (!skip_transition)
+      transition(m_menu_stack.empty() ? nullptr : m_menu_stack.back().get(), menu.get());
+    m_menu_stack.clear();
+    m_menu_stack.push_back(std::move(menu));
+  }
+  else
+  {
+    if (!skip_transition)
+      transition<Menu, Menu>(m_menu_stack.empty() ? nullptr : m_menu_stack.back().get(), nullptr);
+    m_menu_stack.clear();
+  }
+
+  // just to be sure...
+  InputManager::current()->reset();
 }
 
 void
@@ -307,17 +245,11 @@ MenuManager::push_menu(int id, bool skip_transition)
 }
 
 void
-MenuManager::set_menu(int id)
-{
-  set_menu(MenuStorage::instance().create(static_cast<MenuStorage::MenuId>(id)));
-}
-
-void
 MenuManager::push_menu(std::unique_ptr<Menu> menu, bool skip_transition)
 {
   assert(menu);
-  if (!skip_transition) transition(m_menu_stack.empty() ? nullptr : m_menu_stack.back().get(),
-                         menu.get());
+  if (!skip_transition)
+    transition(m_menu_stack.empty() ? nullptr : m_menu_stack.back().get(), menu.get());
   m_menu_stack.push_back(std::move(menu));
 }
 
@@ -327,212 +259,60 @@ MenuManager::pop_menu(bool skip_transition)
   if (m_menu_stack.empty())
   {
     log_warning << "trying to pop on an empty menu_stack" << std::endl;
+    return;
   }
-  else
-  {
-    if (!skip_transition) transition(m_menu_stack.back().get(),
-                           (m_menu_stack.size() >= 2)
-                           ? m_menu_stack[m_menu_stack.size() - 2].get()
-                           : nullptr);
 
-    m_menu_stack.pop_back();
-  }
+  if (!skip_transition)
+    transition(m_menu_stack.back().get(),
+               m_menu_stack.size() >= 2 ? m_menu_stack[m_menu_stack.size() - 2].get() : nullptr);
+  m_menu_stack.pop_back();
 }
 
 void
-MenuManager::set_menu(std::unique_ptr<Menu> menu)
+MenuManager::clear_menu_stack(bool skip_transition)
 {
-  if (menu)
-  {
-    transition(m_menu_stack.empty() ? nullptr : m_menu_stack.back().get(),
-               menu.get());
-    m_menu_stack.clear();
-    m_menu_stack.push_back(std::move(menu));
-  }
-  else
-  {
-    transition(m_menu_stack.empty() ? nullptr : m_menu_stack.back().get(),
-               nullptr, true);
-    m_menu_stack.clear();
-  }
-
-  // just to be sure...
-  InputManager::current()->reset();
-}
-
-void
-MenuManager::clear_menu_stack()
-{
-  transition(m_menu_stack.empty() ? nullptr : m_menu_stack.back().get(),
-             nullptr, true);
+  if (!skip_transition)
+    transition<Menu, Menu>(m_menu_stack.empty() ? nullptr : m_menu_stack.back().get(), nullptr);
   m_menu_stack.clear();
 }
+
 
 void
 MenuManager::on_window_resize()
 {
   for (const auto& menu : m_menu_stack)
-  {
     menu->on_window_resize();
-  }
+}
+
+
+bool
+MenuManager::is_active() const
+{
+  return !m_menu_stack.empty();
+}
+
+bool
+MenuManager::has_dialog() const
+{
+  return m_dialog.current || m_dialog.has_next;
 }
 
 Menu*
 MenuManager::current_menu() const
 {
   if (m_menu_stack.empty())
-  {
     return nullptr;
-  }
   else
-  {
     return m_menu_stack.back().get();
-  }
 }
 
 Menu*
 MenuManager::previous_menu() const
 {
   if (m_menu_stack.size() < 2)
-  {
     return nullptr;
-  }
   else
-  {
     return m_menu_stack.end()[-2].get();
-  }
-}
-
-void
-MenuManager::transition(Menu* from, Menu* to, bool call_this) // "call_this" -> calls this specific overload to prevent ambiguous calls
-{
-  if (!from && !to)
-  {
-    return;
-  }
-  else
-  {
-    Rectf from_rect;
-    if (from)
-    {
-      from_rect = menu2rect(*from);
-    }
-    else
-    {
-      from_rect = Rectf(to->get_center_pos(), Sizef(0, 0));
-    }
-
-    Rectf to_rect;
-    if (to)
-    {
-      to_rect = menu2rect(*to);
-    }
-    else
-    {
-      to_rect = Rectf(from->get_center_pos(), Sizef(0, 0));
-    }
-
-    m_transition->start(from_rect, to_rect);
-  }
-}
-
-void
-MenuManager::transition(Menu* from, Dialog* to)
-{
-  if (!from && !to)
-  {
-    return;
-  }
-  else
-  {
-    Rectf from_rect;
-    if (from)
-    {
-      from_rect = menu2rect(*from);
-    }
-    else
-    {
-      from_rect = Rectf(to->get_center_pos(), Sizef(0, 0));
-    }
-
-    Rectf to_rect;
-    if (to)
-    {
-      to_rect = dialog2rect(*to);
-    }
-    else
-    {
-      to_rect = Rectf(from->get_center_pos(), Sizef(0, 0));
-    }
-
-    m_transition->start(from_rect, to_rect);
-  }
-}
-
-void
-MenuManager::transition(Dialog* from, Menu* to)
-{
-  if (!from && !to)
-  {
-    return;
-  }
-  else
-  {
-    Rectf from_rect;
-    if (from)
-    {
-      from_rect = dialog2rect(*from);
-    }
-    else
-    {
-      from_rect = Rectf(to->get_center_pos(), Sizef(0, 0));
-    }
-
-    Rectf to_rect;
-    if (to)
-    {
-      to_rect = menu2rect(*to);
-    }
-    else
-    {
-      to_rect = Rectf(from->get_center_pos(), Sizef(0, 0));
-    }
-
-    m_transition->start(from_rect, to_rect);
-  }
-}
-
-void
-MenuManager::transition(Dialog* from, Dialog* to)
-{
-  if (!from && !to)
-  {
-    return;
-  }
-  else
-  {
-    Rectf from_rect;
-    if (from)
-    {
-      from_rect = dialog2rect(*from);
-    }
-    else
-    {
-      from_rect = Rectf(to->get_center_pos(), Sizef(0, 0));
-    }
-
-    Rectf to_rect;
-    if (to)
-    {
-      to_rect = dialog2rect(*to);
-    }
-    else
-    {
-      to_rect = Rectf(from->get_center_pos(), Sizef(0, 0));
-    }
-
-    m_transition->start(from_rect, to_rect);
-  }
 }
 
 /* EOF */
