@@ -172,11 +172,10 @@ void
 WorldMap::on_escape_press()
 {
   // Show or hide the menu
-  if (!MenuManager::instance().is_active()) {
+  if (!MenuManager::instance().is_active())
+  {
     MenuManager::instance().set_menu(MenuStorage::WORLDMAP_MENU);
     m_tux->set_direction(Direction::NONE);  // stop tux movement when menu is called
-  } else {
-    MenuManager::instance().clear_menu_stack();
   }
 }
 
